Update `VehiclePathGrid.cs`.
The value returned by `thingGrid.ThingListAt()` is a raw list that can mutate unexpectedly. Mutation of a list invalidates any enumerators of the list, leading to exceptions. This fixes an issue that frequently happens in my game, presumably it is a soft incompatibility with PUAH, but I don't have hard evidence, just an empirical guess based on a similar problem in other mods what were traced back to it.